requirements.txt
numpy==1.26.4
tensorflow==2.16.2
Pillow==11.3.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
tkinterdnd2==0.4.3
opencv-python==4.12.0.88
